% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prelabelled.R
\name{is.prelabelled}
\alias{is.prelabelled}
\title{Test if a vector is prelabelled}
\usage{
is.prelabelled(x)
}
\arguments{
\item{x}{An object.}
}
\value{
Logical scalar.
}
\description{
Determine whether an object inherits from the
\code{"prelabelled"} class.
}
\details{
Useful for lightweight semantic stabilization and contextual
reconstruction workflows where observational evidence and
semantic assertions remain explicitly separated.
}
\seealso{
\code{\link[=prelabel]{prelabel()}},
\code{\link[dataset:is.prelabelled]{dataset::is.prelabelled()}}
}
