% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalize_git_remote}
\alias{normalize_git_remote}
\title{Normalize Git remote identifiers}
\usage{
normalize_git_remote(x)
}
\arguments{
\item{x}{Character vector of Git remotes.}
}
\value{
Character vector of normalized Git repository identifiers.
}
\description{
Converts Git remotes to a stable comparable form.
}
\details{
Normalization currently:
\itemize{
\item converts GitHub SSH remotes to HTTPS form
\item removes trailing \code{.git}
\item removes trailing \code{/}
}
}
\keyword{internal}
