% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recordset_df.R
\name{new_recordset_df}
\alias{new_recordset_df}
\title{Internal constructor for recordset_df}
\usage{
new_recordset_df(x, record_identifier = NULL, record_part_identifier = NULL)
}
\arguments{
\item{x}{A dataset_df object.}

\item{record_identifier}{Optional name of the column containing
Record identifiers.}

\item{record_part_identifier}{Optional name of the column containing
Record Part identifiers.}

\item{record_set_identifier}{Optional identifier of the Record Set.}
}
\value{
A recordset_df object inheriting from dataset_df.
}
\description{
Low-level constructor for recordset_df objects.

This function extends an existing dataset_df with lightweight
Record Set semantics by:
\itemize{
\item assigning the recordset_df class;
\item optionally assigning a Record Set identifier and provenance;
\item optionally declaring Record and Record Part identifier columns as
rico:Identifier using \code{\link[dataset:defined]{dataset::defined()}}.
}

Unlike \code{\link[=recordset_df]{recordset_df()}}, this constructor assumes that dataset-level
metadata have already been created and performs no coercion from
ordinary data.frame objects.
}
\keyword{internal}
